👌(courses) update organization_plugin following review
Update to the organization_plugin html to show:
 -published org only on published page.
 -published org on draft page unless there is
  no published org in which case show the draft org
This corresponds to how organizations show up on course pages.

Streamline the style file for organization plugin.

Update the tests related to organization plugin

Change the related_name of organization plugin model to correspond
to how foreign key relate to other object.